requirements.txt
praw
pandas
python-dotenv